+++ src/elements/Code.php
@@ -4,26 +4,33 @@ namespace verbb\giftvoucher\elements;
 use verbb\giftvoucher\GiftVoucher;
 use verbb\giftvoucher\elements\db\CodeQuery;
 use verbb\giftvoucher\events\GenerateCodeEvent;
-use verbb\giftvoucher\models\VoucherTypeModel;
 use verbb\giftvoucher\records\CodeRecord;
-
 use Craft;
 use craft\base\Element;
 use craft\db\Query;
-use craft\elements\db\ElementQuery;
 use craft\elements\db\ElementQueryInterface;
 use craft\elements\actions\Delete;
 use craft\helpers\ArrayHelper;
 use craft\helpers\DateTimeHelper;
 use craft\helpers\UrlHelper;
 use craft\validators\DateTimeValidator;
-
 use craft\commerce\Plugin as Commerce;
 use craft\commerce\elements\Order;
-
-use Exception;
 use yii\base\InvalidConfigException;
 
+/**
+ * Class Code
+ * @package verbb\giftvoucher\elements
+ *
+ * @property mixed  $amount
+ * @property string $voucherName
+ * @property null   $lineItem
+ * @property mixed  $redemptions
+ * @property null   $voucherType
+ * @property null   $voucher
+ * @property mixed  $name
+ * @property null   $order
+ */
 class Code extends Element
 {
     // Constants
@@ -117,6 +124,11 @@ class Code extends Element
         return $this->currentAmount;
     }
 
+    public function getFieldLayout()
+    {
+        return parent::getFieldLayout() ?? GiftVoucher::getInstance()->getSettings()->getFieldLayout();
+    }
+
     public function getCpEditUrl(): string
     {
         return UrlHelper::cpUrl('gift-voucher/codes/' . $this->id);
@@ -191,6 +203,16 @@ class Code extends Element
         ]);
 
         return $actions;
+    }
+
+    /**
+     * Codes can now have content as well
+     *
+     * @return bool
+     */
+    public static function hasContent(): bool
+    {
+        return true;
     }
 
     public static function eagerLoadingMap(array $sourceElements, string $handle)
@@ -287,7 +309,7 @@ class Code extends Element
             // set the codeKey to the Code as well to use it directly
             $this->codeKey = $codeRecord->codeKey;
         }
-        
+
         $codeRecord->originalAmount = $this->originalAmount;
         $codeRecord->currentAmount = $this->currentAmount;
         $codeRecord->expiryDate = $this->expiryDate;
@@ -304,6 +326,8 @@ class Code extends Element
         }
 
         $codeRecord->save(false);
+
+        parent::afterSave($isNew);
     }