+++ src/services/CodesService.php
@@ -1,23 +1,29 @@
 <?php
 namespace verbb\giftvoucher\services;
 
+use craft\base\Element;
+use craft\helpers\ArrayHelper;
+use verbb\giftvoucher\events\PopulateCodeFromLineItemEvent;
 use verbb\giftvoucher\GiftVoucher;
 use verbb\giftvoucher\adjusters\GiftVoucherAdjuster;
 use verbb\giftvoucher\elements\Code;
 use verbb\giftvoucher\elements\Voucher;
 use verbb\giftvoucher\models\RedemptionModel;
-
 use Craft;
-use craft\events\UserEvent;
-
 use craft\commerce\elements\Order;
 use craft\commerce\models\LineItem;
-
 use yii\base\Component;
 use yii\base\Event;
+use yii\base\ModelEvent;
 
 class CodesService extends Component
 {
+    /**
+     * This event is fired when a new Code is created after an order is complete
+     * to give users a chance to modify the field layout
+     */
+    const EVENT_POPULATE_CODE_FROM_LINE_ITEM = 'populateCodeFromLineItem';
+
     // Public Methods
     // =========================================================================
 
@@ -28,6 +34,7 @@ class CodesService extends Component
 
     public static function handleCompletedOrder(Event $event)
     {
+        /** @var Order $order */
         $order = $event->sender;
         $lineItems = $order->getLineItems();
 
@@ -75,6 +82,18 @@ class CodesService extends Component
         }
     }
 
+    /**
+     * Create a Code after an Order is completed
+     *
+     * @param \verbb\giftvoucher\elements\Voucher $voucher
+     * @param \craft\commerce\elements\Order      $order
+     * @param \craft\commerce\models\LineItem     $lineItem
+     *
+     * @throws \Throwable
+     * @throws \craft\errors\ElementNotFoundException
+     * @throws \yii\base\Exception
+     * @return bool
+     */
     public function codeVoucherByOrder(Voucher $voucher, Order $order, LineItem $lineItem): bool
     {
         $code = new Code();
@@ -85,7 +104,98 @@ class CodesService extends Component
         $code->originalAmount = $lineItem->price;
         $code->currentAmount = $lineItem->price;
 
+        // add the field layout fields
+        $customFields = $this->populateCodeByLineItem($code, $lineItem);
+
+        // give plugins a chance to change/modify it
+        if ($this->hasEventHandlers(self::EVENT_POPULATE_CODE_FROM_LINE_ITEM)) {
+            $this->trigger(self::EVENT_POPULATE_CODE_FROM_LINE_ITEM, new PopulateCodeFromLineItemEvent([
+                'code'          => $code,
+                'order'         => $order,
+                'lineItem'      => $lineItem,
+                'customFields'  => $customFields,
+                'voucher'       => $voucher
+            ]));
+        }
+
+        // TODO: Maybe set the scenario to live to validate the custom fields
+        // we already validate the line item before it can be added to the cart -> they are fine the moment
+        // they are added but what's the proper scenario if a client changes the required fields between creating a
+        // line item and completing the order ¯\_(ツ)_/¯
+        // $code->setScenario(Element::SCENARIO_LIVE);
+
         return Craft::$app->getElements()->saveElement($code);
+    }
+
+    /**
+     * Populates a Code by LineItem options and return the valid custom fields
+     *
+     * @param \verbb\giftvoucher\elements\Code $code
+     * @param \craft\commerce\models\LineItem  $lineItem
+     *
+     * @return array
+     */
+    public function populateCodeByLineItem(Code $code, LineItem $lineItem): array
+    {
+        $settings = GiftVoucher::getInstance()->getSettings();
+        $fieldsPath = $settings->fieldsPath;
+        $fieldLayoutId = $settings->fieldLayoutId;
+        $validFields = [];
+
+        if(empty($fieldsPath) === false && $settings->fieldLayoutId !== null){
+            // set the field layout id
+            $code->fieldLayoutId = $fieldLayoutId;
+            // grab the options from the lineItems, those may contain the field values
+            $options = $lineItem->getOptions();
+            // get the correct path specified by users
+            $customFields = ArrayHelper::getValue($options, $fieldsPath, []);
+
+
+            // okay that might seems a little bit creepy but imagine the case the field layout changes
+            // between storing the line item and creating the code or if the user changes the `fieldsPath` setting
+            // we need to make sure the fields inserted in the options contain valid, still existing fields
+            // and only if that's the case we want to set them otherwise users will see exceptions
+            // that's why we loop every valid field in the layout
+            $fieldLayout = $code->getFieldLayout();
+            if($fieldLayout !== null && ($fields = $fieldLayout->getFields())){
+                /** @var \craft\base\Field $field */
+                foreach ($fields as $field){
+                    $fieldHandle = $field->handle;
+                    if(isset($customFields[$fieldHandle])){
+                        $code->setFieldValue($fieldHandle, $customFields[$fieldHandle]);
+                        $validFields[$fieldHandle] = $customFields[$fieldHandle];
+                    }
+                }
+            }
+        }
+
+        return $validFields;
+    }
+
+    /**
+     * Validate Line Items that are Vouchers based on required Field Layout Fields
+     *
+     * @param \yii\base\ModelEvent $event
+     */
+    public function handleValidateLineItem(ModelEvent $event)
+    {
+        /** @var LineItem $lineItem */
+        $lineItem = $event->sender;
+        $purchasable = $lineItem->getPurchasable();
+        // make sure it's a Voucher
+        if($purchasable instanceof Voucher){
+            // try to create a new Code based on the LineItem
+            $code = new Code();
+            $code->voucherId = $purchasable->id;
+            $this->populateCodeByLineItem($code, $lineItem);
+            $code->setScenario(Element::SCENARIO_LIVE);
+            // if validation fails it means you can't create a code with the current settings
+            if($code->validate() === false){
+                // invalidate it -> let users know about missing/wrong fields
+                $lineItem->addErrors($code->getErrors());
+                $event->isValid = false;
+            }
+        }
     }
 
     public function generateCodeKey(): string